requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==1.5.0
blis==0.4.1
boto3==1.12.26
botocore==1.15.26
Bottleneck==1.3.2
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
construct==2.10.56
cycler==0.10.0
cymem==2.0.3
Cython==0.29.17
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
dominate==2.5.1
entrypoints==0.3
fastai==1.0.60
fastprogress==0.2.3
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
Hyphenate==1.1.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
midi2audio==0.1.1
mido==1.2.9
mistune==0.8.4
more-itertools==8.2.0
murmurhash==1.0.2
music21==5.7.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.14.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
pretty-midi==0.2.8
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
py-midi==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
Pygments==2.6.1
pynvx==1.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
pyserial==3.4
pysndfx==0.3.6
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
smart-open==1.10.0
soupsieve==2.0
spacy==2.2.4
SQLAlchemy==1.3.9
srsly==1.0.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
utils==1.0.1
visitor==0.1.3
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
